`AccountServiceSpec` and `SetSavingsGoalSpec` are now implemented in terms of a `TestF` alias for a type constructor which is used to instantiate the services being tested, as well as the fakes to support the testing. The requirement on `TestF` is that there should be a `MonadError` for it, as this is what `AccountServiceImpl` requires.
`TestF` is now aliased to `Try` instead of `Future`, which removes all of the dependency on futures and timeouts and so forth. One of the advantages of this is that if you use a debugger to step through a test you don't have to worry about an `await` timing out while you are inspecting things.