I've added a `.scalafmt.conf` file, based on the file from `http-verbs`, and re-formatted the code. The headers plugin has updated the copyright year to 2019 in all files.
I've also disabled `-Ywarn-unused-import` and enabled `-Xfatal-warnings` and fixed up the return types in the it test wiremock stub generators to not warn about discarded non-unit values. We can't have both `-Ywarn-unused-import` and `-Xfatal-warnings` enabled at the same time because of the unused imports in the source that Play creates for the routes. On balance, fatal-warnings is probably more useful to us.