[DC] Remove workaround `health2.routes`
When I originally did the static wiring for the components there was a problem with the version of the health routes from the play 2.5 version of the `play-health` library. Because it had been built with the static routes compiler, the `health.Routes` class was using the guice injector to wire things, which was not compatible with the static DI, so I added a `health2.routes` file that replicated the routes using the non-static routes compiler.

Now that we've upgraded to Play 2.6, we're using a version of `play-health` that has a non-statically-compiled routes file (i.e. `health.Routes` is now a class that takes its dependencies as parameters) so I have removed the work-around `health2.routes` file and updated the `prod.routes`.